Refactor UMKM data into markdown content
