+++ content/umkm/stores/01-nusantara-coffee.md
@@ -1,0 +1,33 @@
+---
+order: 1
+id: nusantara-coffee
+name: Nusantara Coffee Roasters
+category: Kopi Spesialti
+location: Bandung, Jawa Barat
+whatsappNumber: '6281234567001'
+tagline: >-
+  Mempersembahkan biji kopi single origin terbaik dari seluruh Nusantara.
+heroImage: https://images.unsplash.com/photo-1507133750040-4a8f57021571?auto=format&fit=crop&w=800&q=80
+description: >-
+  Nusantara Coffee Roasters bekerja langsung dengan petani kopi Indonesia untuk menghadirkan racikan biji kopi spesialti dengan profil rasa khas wilayah. Sangat cocok untuk para penikmat kopi manual brew maupun espresso.
+highlights:
+  - Kemitraan langsung dengan petani lokal
+  - Pengiriman biji kopi fresh roast setiap minggu
+  - Kelas edukasi brewing gratis setiap akhir pekan
+products:
+  - name: Kopi Gayo Honey Process
+    price: 'Rp120.000 / 200gr'
+    description: >-
+      Biji kopi Arabika dari Aceh dengan proses honey yang menghasilkan rasa manis madu dan karamel.
+    image: https://images.unsplash.com/photo-1447933601403-0c6688de566e?auto=format&fit=crop&w=800&q=80
+  - name: Blend Espresso Nusantara
+    price: 'Rp150.000 / 500gr'
+    description: >-
+      Perpaduan biji kopi Toraja, Flores, dan Java untuk menciptakan espresso dengan body penuh dan aroma cokelat.
+    image: https://images.unsplash.com/photo-1470337458703-46ad1756a187?auto=format&fit=crop&w=800&q=80
+  - name: Cold Brew Ready to Drink
+    price: 'Rp35.000 / botol'
+    description: >-
+      Cold brew 100% Arabika dengan rasa lembut dan rendah asam, dikemas praktis siap minum.
+    image: https://images.unsplash.com/photo-1447933601403-0c6688de566e?auto=format&fit=crop&w=800&q=80
+---
