+++ content/umkm/stores/02-batik-pradana.md
@@ -1,0 +1,33 @@
+---
+order: 2
+id: batik-pradana
+name: Batik Pradana
+category: Batik Tulis
+location: Solo, Jawa Tengah
+whatsappNumber: '6281234567002'
+tagline: >-
+  Menghidupkan warisan budaya melalui batik tulis kontemporer.
+heroImage: https://images.unsplash.com/photo-1591769225440-811ad7d9e0d1?auto=format&fit=crop&w=800&q=80
+description: >-
+  Batik Pradana memadukan motif klasik dengan sentuhan modern sehingga cocok dikenakan pada acara formal maupun kasual. Semua kain diproduksi secara etis oleh pengrajin lokal.
+highlights:
+  - Menggunakan pewarna alami ramah lingkungan
+  - Motif eksklusif edisi terbatas
+  - Custom order untuk seragam kantor dan komunitas
+products:
+  - name: Kain Batik Sekar Jagad
+    price: Rp750.000
+    description: >-
+      Kain batik tulis motif Sekar Jagad dengan kombinasi warna indigo dan sogan yang elegan.
+    image: https://images.unsplash.com/photo-1512436991641-6745cdb1723f?auto=format&fit=crop&w=800&q=80
+  - name: Outer Batik Larasati
+    price: Rp450.000
+    description: >-
+      Outer wanita lengan panjang dengan potongan modern dan motif Larasati yang lembut.
+    image: https://images.unsplash.com/photo-1556306535-0c0946175acc?auto=format&fit=crop&w=800&q=80
+  - name: Kemeja Batik Pria Parang
+    price: Rp520.000
+    description: >-
+      Kemeja pria lengan pendek dengan motif Parang Rusak kontemporer.
+    image: https://images.unsplash.com/photo-1617032212157-5563d9739dc6?auto=format&fit=crop&w=800&q=80
+---
