+++ content/umkm/stores/03-rempah-ceria.md
@@ -1,0 +1,33 @@
+---
+order: 3
+id: rempah-ceria
+name: Rempah Ceria
+category: Bumbu Masak
+location: Yogyakarta, DI Yogyakarta
+whatsappNumber: '6281234567003'
+tagline: >-
+  Bumbu instan rumahan dengan rasa autentik Indonesia.
+heroImage: https://images.unsplash.com/photo-1556906781-9a412961c28c?auto=format&fit=crop&w=800&q=80
+description: >-
+  Rempah Ceria menghadirkan bumbu instan siap pakai tanpa bahan pengawet, dibuat dari rempah segar pilihan dan resep turun temurun.
+highlights:
+  - Tanpa MSG dan pengawet buatan
+  - Kemasan praktis dengan porsi pas untuk keluarga
+  - Varian lengkap masakan Nusantara
+products:
+  - name: Bumbu Rendang Padang
+    price: Rp28.000
+    description: >-
+      Bumbu siap pakai dengan rempah lengkap untuk memasak rendang daging khas Minang.
+    image: https://images.unsplash.com/photo-1590725121839-ff2c1f4423fe?auto=format&fit=crop&w=800&q=80
+  - name: Bumbu Ayam Woku
+    price: Rp25.000
+    description: >-
+      Racikan rempah woku khas Manado dengan cita rasa pedas segar.
+    image: https://images.unsplash.com/photo-1562967914-608f82629710?auto=format&fit=crop&w=800&q=80
+  - name: Sambal Matah Bali
+    price: Rp22.000
+    description: >-
+      Sambal matah segar dengan serai dan jeruk limau, cocok untuk lauk goreng atau bakar.
+    image: https://images.unsplash.com/photo-1528832983572-5cc2c3df96de?auto=format&fit=crop&w=800&q=80
+---
