+++ content/umkm/stores/04-kerajinan-rumah-bambu.md
@@ -1,0 +1,33 @@
+---
+order: 4
+id: kerajinan-rumah-bambu
+name: Kerajinan Rumah Bambu
+category: Kerajinan Tangan
+location: Bogor, Jawa Barat
+whatsappNumber: '6281234567004'
+tagline: >-
+  Dekorasi rumah ramah lingkungan berbahan bambu pilihan.
+heroImage: https://images.unsplash.com/photo-1489515217757-5fd1be406fef?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kerajinan Rumah Bambu menciptakan furnitur dan dekorasi rumah minimalis yang terbuat dari bambu berkualitas tinggi. Produk dirakit oleh pengrajin yang terampil dengan sentuhan modern.
+highlights:
+  - Bambu dipanen berkelanjutan dari hutan rakyat
+  - Finishing tahan rayap dan cuaca
+  - Desain minimalis cocok untuk interior modern
+products:
+  - name: Lampu Gantung Bambu Lontar
+    price: Rp350.000
+    description: >-
+      Lampu gantung dengan anyaman bambu motif lontar yang memberikan cahaya hangat alami.
+    image: https://images.unsplash.com/photo-1519710164239-da123dc03ef4?auto=format&fit=crop&w=800&q=80
+  - name: Kursi Santai Anyaman
+    price: Rp1.250.000
+    description: >-
+      Kursi santai berbahan bambu dengan sandaran anyaman rotan sintetis yang ergonomis.
+    image: https://images.unsplash.com/photo-1519710164239-da123dc03ef4?auto=format&fit=crop&w=800&q=80
+  - name: Rak Dinding Modular
+    price: Rp480.000
+    description: >-
+      Rak dinding bambu dengan modul fleksibel, mudah dipasang dan dikombinasikan.
+    image: https://images.unsplash.com/photo-1489515217757-5fd1be406fef?auto=format&fit=crop&w=800&q=80
+---
