+++ content/umkm/stores/05-kuliner-vegan-hijau.md
@@ -1,0 +1,33 @@
+---
+order: 5
+id: kuliner-vegan-hijau
+name: Kuliner Vegan Hijau
+category: Kuliner Sehat
+location: Jakarta Selatan, DKI Jakarta
+whatsappNumber: '6281234567005'
+tagline: >-
+  Menu plant-based lezat, sehat, dan praktis.
+heroImage: https://images.unsplash.com/photo-1533777419517-3e4017e2e15b?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kuliner Vegan Hijau menghadirkan hidangan plant-based dengan bahan organik lokal. Cocok untuk gaya hidup sehat tanpa mengorbankan rasa.
+highlights:
+  - Menu ganti-ganti tiap minggu
+  - Tanpa bahan pengawet dan gula rafinasi
+  - Konsultasi gizi gratis bersama nutrisionis
+products:
+  - name: Paket Nasi Shirataki Rendang Jamur
+    price: Rp65.000
+    description: >-
+      Rendang jamur tiram dengan bumbu rempah kaya rasa disajikan bersama nasi shirataki rendah kalori.
+    image: https://images.unsplash.com/photo-1485965120184-e220f721d03e?auto=format&fit=crop&w=800&q=80
+  - name: Salad Bowl Tropical Sunshine
+    price: Rp55.000
+    description: >-
+      Salad bowl dengan campuran sayur organik, buah tropis, quinoa, dan dressing jeruk nipis.
+    image: https://images.unsplash.com/photo-1546069901-ba9599a7e63c?auto=format&fit=crop&w=800&q=80
+  - name: Lasagna Vegan Bayam
+    price: Rp75.000
+    description: >-
+      Lasagna panggang dengan saus tomat homemade, keju cashew, dan lapisan bayam organik.
+    image: https://images.unsplash.com/photo-1504674900247-0877df9cc836?auto=format&fit=crop&w=800&q=80
+---
