+++ content/umkm/stores/06-perhiasan-perak-cahaya.md
@@ -1,0 +1,33 @@
+---
+order: 6
+id: perhiasan-perak-cahaya
+name: Perhiasan Perak Cahaya
+category: Perhiasan Handmade
+location: Gianyar, Bali
+whatsappNumber: '6281234567006'
+tagline: >-
+  Perhiasan perak handmade dengan sentuhan budaya Bali.
+heroImage: https://images.unsplash.com/photo-1518544801976-3e89f2f1df8e?auto=format&fit=crop&w=800&q=80
+description: >-
+  Perhiasan Perak Cahaya menawarkan koleksi cincin, gelang, dan kalung berbahan perak sterling dengan detail ukir tradisional dan desain modern.
+highlights:
+  - Perak 925 dengan sertifikat keaslian
+  - Desain custom sesuai permintaan
+  - Pengemasan eksklusif untuk hadiah
+products:
+  - name: Cincin Lotus Harmony
+    price: Rp450.000
+    description: >-
+      Cincin perak 925 dengan ukiran motif lotus dan aksen batu moonstone.
+    image: https://images.unsplash.com/photo-1522312346375-d1a52e2b99b3?auto=format&fit=crop&w=800&q=80
+  - name: Gelang Bangle Dewi Sri
+    price: Rp680.000
+    description: >-
+      Gelang bangle dengan ukiran padi melambangkan kemakmuran dan keseimbangan.
+    image: https://images.unsplash.com/photo-1518544801976-3e89f2f1df8e?auto=format&fit=crop&w=800&q=80
+  - name: Kalung Pesona Tirta
+    price: Rp520.000
+    description: >-
+      Kalung perak dengan pendant motif ombak dan batu larimar.
+    image: https://images.unsplash.com/photo-1455849318743-b2233052fcff?auto=format&fit=crop&w=800&q=80
+---
