+++ content/umkm/stores/07-kuliner-dodol-garut.md
@@ -1,0 +1,33 @@
+---
+order: 7
+id: kuliner-dodol-garut
+name: Kuliner Dodol Garut
+category: Oleh-oleh Khas
+location: Garut, Jawa Barat
+whatsappNumber: '6281234567007'
+tagline: >-
+  Dodol tradisional dengan inovasi rasa kekinian.
+heroImage: https://images.unsplash.com/photo-1551024709-8f23befc6f87?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kuliner Dodol Garut menghadirkan dodol legit khas Garut yang dibuat dari bahan premium dan dimasak perlahan hingga teksturnya lembut.
+highlights:
+  - Tanpa pemanis buatan dan pengawet
+  - Varian rasa modern seperti matcha dan kopi
+  - Tersedia paket hampers eksklusif
+products:
+  - name: Dodol Garut Original
+    price: Rp30.000
+    description: >-
+      Dodol klasik dengan rasa gula aren dan santan yang kaya.
+    image: https://images.unsplash.com/photo-1504674900247-0877df9cc836?auto=format&fit=crop&w=800&q=80
+  - name: Dodol Matcha Almond
+    price: Rp38.000
+    description: >-
+      Paduan dodol lembut dengan bubuk matcha Jepang dan taburan almond panggang.
+    image: https://images.unsplash.com/photo-1572441710534-68094d3e8553?auto=format&fit=crop&w=800&q=80
+  - name: Dodol Kopi Susu
+    price: Rp35.000
+    description: >-
+      Dodol rasa kopi susu dengan aroma kopi nusantara dan tekstur chewy.
+    image: https://images.unsplash.com/photo-1466978913421-dad2ebd01d17?auto=format&fit=crop&w=800&q=80
+---
