+++ content/umkm/stores/08-eco-laundry.md
@@ -1,0 +1,33 @@
+---
+order: 8
+id: eco-laundry
+name: Eco Laundry
+category: Jasa Kebersihan
+location: Surabaya, Jawa Timur
+whatsappNumber: '6281234567008'
+tagline: >-
+  Layanan laundry ramah lingkungan dengan deterjen biodegradable.
+heroImage: https://images.unsplash.com/photo-1527515637462-cff94eecc1ac?auto=format&fit=crop&w=800&q=80
+description: >-
+  Eco Laundry menggunakan deterjen bebas fosfat dan teknologi hemat air untuk merawat pakaian pelanggan. Tersedia layanan antar jemput area Surabaya.
+highlights:
+  - Menggunakan energi panel surya di workshop
+  - Layanan same day untuk pelanggan prioritas
+  - Program loyalty dengan poin eco reward
+products:
+  - name: Paket Laundry Kiloan Eco Wash
+    price: Rp12.000 / kg
+    description: >-
+      Layanan cuci lipat menggunakan deterjen eco-friendly dan pewangi alami.
+    image: https://images.unsplash.com/photo-1522335789203-aabd1fc54bc9?auto=format&fit=crop&w=800&q=80
+  - name: Paket Premium Dry Clean
+    price: Rp45.000 / item
+    description: >-
+      Dry clean ramah lingkungan untuk jas, kebaya, dan gaun pesta.
+    image: https://images.unsplash.com/photo-1522335789203-aabd1fc54bc9?auto=format&fit=crop&w=800&q=80
+  - name: Cuci Sepatu Sneaker Care
+    price: Rp60.000 / pasang
+    description: >-
+      Perawatan sepatu dengan teknik deep cleaning dan pewangi antibakteri.
+    image: https://images.unsplash.com/photo-1489515217757-5fd1be406fef?auto=format&fit=crop&w=800&q=80
+---
