+++ content/umkm/stores/09-studio-keramik.md
@@ -1,0 +1,33 @@
+---
+order: 9
+id: studio-keramik
+name: Studio Keramik Taman
+category: Kriya Keramik
+location: Bandung, Jawa Barat
+whatsappNumber: '6281234567009'
+tagline: >-
+  Studio keramik artisana dengan sentuhan flora nusantara.
+heroImage: https://images.unsplash.com/photo-1520256862855-398228c41684?auto=format&fit=crop&w=800&q=80
+description: >-
+  Studio Keramik Taman membuat peralatan makan dan dekorasi rumah dari stoneware yang dibakar suhu tinggi sehingga tahan lama dan aman food-grade.
+highlights:
+  - Workshop kelas membuat keramik tiap bulan
+  - Custom set dinnerware untuk kafe dan restoran
+  - Glasir ramah makanan dengan warna earthy
+products:
+  - name: Set Cangkir Kopi Flora
+    price: Rp320.000 / set
+    description: >-
+      Dua cangkir espresso dengan piring kecil motif daun tropis.
+    image: https://images.unsplash.com/photo-1522684462855-b97ff5949c4a?auto=format&fit=crop&w=800&q=80
+  - name: Mangkok Salad Teraso
+    price: Rp280.000
+    description: >-
+      Mangkok salad berdiameter 20 cm dengan tekstur teraso alami.
+    image: https://images.unsplash.com/photo-1489515217757-5fd1be406fef?auto=format&fit=crop&w=800&q=80
+  - name: Vas Bunga Kendi Rimba
+    price: Rp450.000
+    description: >-
+      Vas bunga tinggi dengan motif ukir hutan hujan dan finishing matte.
+    image: https://images.unsplash.com/photo-1446292532430-3e76f6ab6444?auto=format&fit=crop&w=800&q=80
+---
