+++ content/umkm/stores/10-aroma-therapy.md
@@ -1,0 +1,33 @@
+---
+order: 10
+id: aroma-therapy
+name: Aroma Therapy Nusantara
+category: Aromaterapi
+location: Denpasar, Bali
+whatsappNumber: '6281234567010'
+tagline: >-
+  Minyak aromaterapi dan lilin kedelai dengan aroma khas nusantara.
+heroImage: https://images.unsplash.com/photo-1515377905703-c4788e51af15?auto=format&fit=crop&w=800&q=80
+description: >-
+  Aroma Therapy Nusantara meracik minyak atsiri dan lilin aromaterapi dari bahan alami, terinspirasi dari rempah dan flora Indonesia.
+highlights:
+  - Menggunakan minyak esensial murni
+  - Kemasan ramah lingkungan dan dapat didaur ulang
+  - Memberdayakan petani rempah lokal
+products:
+  - name: Minyak Esensial Boreh Bali
+    price: Rp140.000
+    description: >-
+      Campuran minyak serai, cengkeh, dan jahe untuk relaksasi otot.
+    image: https://images.unsplash.com/photo-1493663284031-b7e3aefcae8e?auto=format&fit=crop&w=800&q=80
+  - name: Lilin Aromaterapi Senja Nusantara
+    price: Rp185.000
+    description: >-
+      Lilin kedelai dengan aroma kayu manis, pala, dan vanilla.
+    image: https://images.unsplash.com/photo-1489515217757-5fd1be406fef?auto=format&fit=crop&w=800&q=80
+  - name: Reed Diffuser Hutan Tropis
+    price: Rp210.000
+    description: >-
+      Diffuser aroma hutan hujan dengan notes kayu dan citrus.
+    image: https://images.unsplash.com/photo-1522335789203-aabd1fc54bc9?auto=format&fit=crop&w=800&q=80
+---
