+++ content/umkm/stores/11-tekstil-tenun.md
@@ -1,0 +1,33 @@
+---
+order: 11
+id: tekstil-tenun
+name: Tekstil Tenun Lestari
+category: Tekstil Tradisional
+location: Kupang, Nusa Tenggara Timur
+whatsappNumber: '6281234567011'
+tagline: >-
+  Kain tenun ikat eksklusif yang memadukan motif adat dengan gaya modern.
+heroImage: https://images.unsplash.com/photo-1542293787938-4d2226cba1f2?auto=format&fit=crop&w=800&q=80
+description: >-
+  Tekstil Tenun Lestari bekerja sama dengan penenun perempuan NTT untuk menghadirkan kain ikat berkualitas tinggi dengan pewarna alami.
+highlights:
+  - Mendukung keberlanjutan komunitas penenun lokal
+  - Pewarna alami dari tumbuhan endemik
+  - Koleksi terbatas dengan sertifikat keaslian
+products:
+  - name: Kain Tenun Ikat Motif Lontar
+    price: Rp1.800.000
+    description: >-
+      Kain panjang 2,5 meter dengan motif lontar khas Timor dan warna alam.
+    image: https://images.unsplash.com/photo-1542293787938-4d2226cba1f2?auto=format&fit=crop&w=800&q=80
+  - name: Selendang Ikat Mawar Sumba
+    price: Rp650.000
+    description: >-
+      Selendang lembut dengan motif bunga mawar dan warna merah marun.
+    image: https://images.unsplash.com/photo-1617032212157-5563d9739dc6?auto=format&fit=crop&w=800&q=80
+  - name: Kemeja Tenun Modern
+    price: Rp980.000
+    description: >-
+      Kemeja pria dengan kombinasi kain tenun ikat dan katun premium.
+    image: https://images.unsplash.com/photo-1512436991641-6745cdb1723f?auto=format&fit=crop&w=800&q=80
+---
