+++ content/umkm/stores/12-kuliner-madu.md
@@ -1,0 +1,33 @@
+---
+order: 12
+id: kuliner-madu
+name: Kuliner Madu Nusantara
+category: Produk Alami
+location: Malang, Jawa Timur
+whatsappNumber: '6281234567012'
+tagline: >-
+  Madu murni dan olahan propolis dari peternak lokal.
+heroImage: https://images.unsplash.com/photo-1505576399279-565b52d4ac71?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kuliner Madu Nusantara mengkurasi madu raw, propolis, dan olahan kesehatan lainnya langsung dari peternak lebah Indonesia.
+highlights:
+  - Madu diuji laboratorium untuk memastikan kemurnian
+  - Edukasi budidaya lebah bagi komunitas
+  - Kemasan food grade yang aman dikirim jarak jauh
+products:
+  - name: Madu Hutan Mentah 500ml
+    price: Rp120.000
+    description: >-
+      Madu hutan dengan rasa floral kuat dan tekstur kental.
+    image: https://images.unsplash.com/photo-1505576399279-565b52d4ac71?auto=format&fit=crop&w=800&q=80
+  - name: Propolis Spray Herbal
+    price: Rp95.000
+    description: >-
+      Propolis cair dengan kandungan flavonoid tinggi untuk menjaga daya tahan tubuh.
+    image: https://images.unsplash.com/photo-1470337458703-46ad1756a187?auto=format&fit=crop&w=800&q=80
+  - name: Madu Infused Rempah Hangat
+    price: Rp85.000
+    description: >-
+      Campuran madu dengan jahe, kayu manis, dan cengkeh siap seduh.
+    image: https://images.unsplash.com/photo-1470337458703-46ad1756a187?auto=format&fit=crop&w=800&q=80
+---
