+++ content/umkm/stores/13-kriya-kulit.md
@@ -1,0 +1,33 @@
+---
+order: 13
+id: kriya-kulit
+name: Kriya Kulit Mandiri
+category: Kriya Kulit
+location: Yogyakarta, DI Yogyakarta
+whatsappNumber: '6281234567013'
+tagline: >-
+  Produk kulit handmade dengan personalisasi inisial.
+heroImage: https://images.unsplash.com/photo-1470337458703-46ad1756a187?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kriya Kulit Mandiri memproduksi aksesori kulit asli seperti dompet, tas, dan strap kamera dengan teknik hand-stitching.
+highlights:
+  - Kulit sapi full grain pilihan
+  - Personalisasi emboss inisial gratis
+  - Garansi perbaikan jahitan seumur hidup
+products:
+  - name: Dompet Lipat Classic Brown
+    price: Rp320.000
+    description: >-
+      Dompet lipat dengan 6 slot kartu dan kompartemen uang terpisah.
+    image: https://images.unsplash.com/photo-1470337458703-46ad1756a187?auto=format&fit=crop&w=800&q=80
+  - name: Tas Selempang Heritage
+    price: Rp980.000
+    description: >-
+      Tas selempang dengan flap magnetik dan ruang untuk tablet 10 inci.
+    image: https://images.unsplash.com/photo-1511389026070-a14ae610a1be?auto=format&fit=crop&w=800&q=80
+  - name: Strap Kamera Traveler
+    price: Rp450.000
+    description: >-
+      Strap kamera kulit dengan bantalan bahu dan pengunci kuningan.
+    image: https://images.unsplash.com/photo-1493663284031-b7e3aefcae8e?auto=format&fit=crop&w=800&q=80
+---
