+++ content/umkm/stores/14-studio-daur-ulang.md
@@ -1,0 +1,33 @@
+---
+order: 14
+id: studio-daur-ulang
+name: Studio Daur Ulang Kreasi
+category: Upcycle Design
+location: Semarang, Jawa Tengah
+whatsappNumber: '6281234567014'
+tagline: >-
+  Produk rumah tangga kreatif dari bahan daur ulang.
+heroImage: https://images.unsplash.com/photo-1466978913421-dad2ebd01d17?auto=format&fit=crop&w=800&q=80
+description: >-
+  Studio Daur Ulang Kreasi mengolah limbah plastik dan tekstil menjadi produk fungsional yang bergaya.
+highlights:
+  - Mengurangi sampah plastik dan tekstil kota
+  - Memberdayakan komunitas bank sampah
+  - Desain custom untuk acara dan korporasi
+products:
+  - name: Tas Totebag Anyam Replas
+    price: Rp180.000
+    description: >-
+      Totebag dari anyaman plastik bekas dengan lapisan kain kanvas.
+    image: https://images.unsplash.com/photo-1522335789203-aabd1fc54bc9?auto=format&fit=crop&w=800&q=80
+  - name: Alas Meja Patchwork Denim
+    price: Rp95.000
+    description: >-
+      Alas meja bundar dari potongan denim bekas yang dijahit patchwork.
+    image: https://images.unsplash.com/photo-1522335789203-aabd1fc54bc9?auto=format&fit=crop&w=800&q=80
+  - name: Pot Tanaman Ecobrick
+    price: Rp65.000
+    description: >-
+      Pot tanaman minimalis dari ecobrick dengan finishing cat ramah lingkungan.
+    image: https://images.unsplash.com/photo-1519710164239-da123dc03ef4?auto=format&fit=crop&w=800&q=80
+---
