+++ content/umkm/stores/15-kopi-susu-botol.md
@@ -1,0 +1,33 @@
+---
+order: 15
+id: kopi-susu-botol
+name: Kopi Susu Botol Harian
+category: Minuman Siap Minum
+location: Tangerang, Banten
+whatsappNumber: '6281234567015'
+tagline: >-
+  Kopi susu siap minum dengan resep favorit generasi urban.
+heroImage: https://images.unsplash.com/photo-1495474472287-4d71bcdd2085?auto=format&fit=crop&w=800&q=80
+description: >-
+  Kopi Susu Botol Harian meracik kopi susu kekinian dengan bahan berkualitas tinggi dan proses cold brew yang menjaga rasa.
+highlights:
+  - Menggunakan susu segar pasteurisasi harian
+  - Tanpa perisa buatan dan pemanis sintetis
+  - Tersedia paket langganan mingguan
+products:
+  - name: Kopi Susu Gula Aren
+    price: Rp25.000
+    description: >-
+      Kopi susu dengan gula aren premium dan espresso blend house blend.
+    image: https://images.unsplash.com/photo-1504753793650-d4a2b783c15e?auto=format&fit=crop&w=800&q=80
+  - name: Kopi Susu Pandan
+    price: Rp27.000
+    description: >-
+      Paduan espresso dengan susu pandan homemade dan aroma wangi daun pandan.
+    image: https://images.unsplash.com/photo-1447933601403-0c6688de566e?auto=format&fit=crop&w=800&q=80
+  - name: Latte Oat Kurma
+    price: Rp30.000
+    description: >-
+      Espresso dengan susu oat dan manis alami dari buah kurma.
+    image: https://images.unsplash.com/photo-1527169402691-feff5539e52c?auto=format&fit=crop&w=800&q=80
+---
